selenium-ingestor/requirements.txt
pandas==2.2.3
selenium==4.27.1
webdriver-manager==4.0.2
beautifulsoup4==4.12.3
google-cloud-storage==2.18.2
